New Feature: filename pattern detection
Infected mails often come with similiar named emails - something "unpaid_invoice_1234.zip" or "report_1234.zip".
This EXPERIMENTAL feature maintains a sqlite database with these filenames and reject files that are similiar named and have previously detected as infected.

The filenames is split into chunks (seperator are _, - and .). The resulting tuple is compared against the other filename entries in the database. If another filenames differ at exactly one chunk, the pattern is recorded in the database.

Example
- weekly_report_1234.zip
- weekly_report-5678.zip
- weekly-report.foobar.zip

will all create the pattern 'weekly-report-[RANDOM]-zip'.

AmavisVT will flag these files as infected, if:
- there are more than "min-filename-patterns" (default: 20) entries with this pattern in the database
- more than "infected-percent" (default: 0.7 = 70%) of these entries have been detected as infected

Additionaly, AmavisVT will send these files to Virustotal if the "auto-report" setting is set to true.

WARNING: This will send the actual content of the attachments to Virustotal. DO NOT enable this setting if
         a legitimate attachment may be caught by the pattern detection!

AmavisVT's database requires sqlite support in python.

The filename pattern detection feature requires the fuzzywuzzy library (https://github.com/seatgeek/fuzzywuzzy) to be installed.